Class Project for Upper division Intermediate Geographical Analysis Course
This project was the class project for an upper division course at UCSB called Intermediate Geographical Data Analysis. The project involved using R to run different statistical analysis on spatial data. I took the framework of this project and applied it to something I found more relatable at the time: Basketball! Most of this code was taken from various lab sessions throughout the year and applied to my project. My understanding of coding at the time was very limited and subsequently, the script I created to run the analysis was very limited but I still had a lot of fun doing this. This course was a chief reason I decided to emphasize in statistics during my undergraduate degree. (The Course Description for this class is as follows: The course builds on students’ prior exposure to probability theory and statistics to develop familiarity with the theory and application of spatial statistics and exploratory spatial data analysis. The primary emphasis in the course is to increase students’ comfort level with finding and testing patterns in spatial data.)